🎨 improve: Mejorando estructura del código modularizando funciones en otro módulos para una mayor organización y se ha resuelto un Bug dentro de la función message_handler que impedía la ejecución de la función si el canal de broma no estaba establecido
