+++ src/utils/handlers/mod.rs
@@ -2,3 +2,4 @@ pub mod sent_messages;
 pub mod deleted_messages;
 pub mod error;
 pub mod edited_messages;
+pub mod misc;
